dodavanje smjestaja preko forme gotovo

+++ src/main/kotlin/hranj/marijan/diplomskirad/repository/KategorijaRepository.kt
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface KategorijaRepository : JpaRepository<Kategorija, Long> {
+interface KategorijaRepository : JpaRepository<Kategorija, Int> {
 
     fun findAllByIdIn(id: List<Int>): Set<Kategorija>
