+++ src/main/kotlin/hranj/marijan/diplomskirad/repository/KorisnikRepository.kt
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface KorisnikRepository : JpaRepository<Korisnik, Long> {
+interface KorisnikRepository : JpaRepository<Korisnik, Int> {
 
     fun findByKorisnickoImeOrEmail(korisnickoIme: String?, email: String?): Korisnik?
