+++ src/main/kotlin/hranj/marijan/diplomskirad/repository/SmjestajRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface SmjestajRepository : JpaRepository<Smjestaj, Long>
+interface SmjestajRepository : JpaRepository<Smjestaj, Int>
